Move DAS storage server factory to das package
